spark: Don't hardcode the number of workers
Before, the Spark test assumed that there would be 4 worker machines.
Because commit 8a69a13 decreased the number of worker machines to 3, the
Spark test failed because of a placement rule that forbade placing two
Spark containers on the same machine.

We now determine the number of Spark workers to deploy based on the
`infrastructure.nWorker` variable.