jenkins: Test Travis-built Quilt binary
Before, the Jenkins job cloned the Quilt repo, and built it locally. We
now download the Quilt binary that Travis builds, and test that.

This will make it easier to test different Quilt releases because we
will simply have to change the binary that we download. Furthermore,
this avoids potential errors where Travis builds a slightly different
binary than Jenkins.