integration-tester: Use Groovy script to post results to Slack
Before this commit, we used a Jenkins plugin to post the build results
to Slack; however, the plugin does not seem to have much activity
anymore, and it lacks some desirable features. This commit removes the
plugin in favor of a Groovy script that directly posts the build results
to Slack. The Groovy script implementation also includes several
enhancements:
- It includes the latest Git commit in the post.
- It includes the Kelda version and number of workers in the post.
- It links failed tests to their output page.
- It pings @channel when tests fail, allowing people to mute the
channel, and only receive notifications when tests fail.